+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 2.0.0~development
 
   * Changes that break retro-compatibility:
-    - 'import thermod' only imports constants and classes useful for clients and monitors
-    - remove 'error_code' from Settings class (it wasn't used anywhere)
+    - `import thermod` only imports constants and classes useful for clients and monitors
+    - remove `error_code` from `Settings` class (it wasn't used anywhere)
+    - move `DEGREE*` variables in `common.py` module
 
   * New features: